Moved the $SERVER_WO_PORT variable up into the area where MySQL parameters are constructed, and created a new variable called $SERVER_PORT to separate out the port.
In my use case, I'd like to set up GUMS with the MySQL database living on
another machine. This seemed to be the best way to handle that without
significantly breaking the existing setup script.

Also relevant, from the MySQL 5.1 reference:
"For connections to localhost, MySQL programs attempt to connect to the local
server by using a Unix socket file.  This occurs even if a --port or -P option
is given to specify a port number. To ensure that the client makes a TCP/IP
connection to the local server, use --host or -h to specify a host name value of
127.0.0.1, or the IP address or name of the local server."
https://dev.mysql.com/doc/refman/5.1/en/connecting.html

I take this to mean that specifying -h and -p won't break connections to
localhost.