Switch VOMSACValidator to a thread-local object.
The VOMSACValidator spawns a helper thread for each instance and re-loads
the CRLs for each new instance.  Hence, it's too heavyweight to do an instance
per validation request.

OTOH, the actual validation is done in a synchronized method, meaning only one
thread can validate at a time per VOMSACValidator object.  So, it's also
inappropriate to be a global object.

Here, I split the difference and have it be a thread-local.